extract the sync initializing logic into a common location
